detection/requirements.txt
torch
torchvision
opencv-python
onnx
onnxruntime-gpu==1.7
numpy